api/requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.12.1
psycopg2-binary
SQLAlchemy==1.4.36
Werkzeug==2.1.2
python-dotenv

